pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = ""
authors = ["Boris Rakovan <boris.rakovan@ataccama.com>"]

[tool.poetry.dependencies]
python = "~3.9"
torch = "^1.11.0"
numpy = "^1.22.3"
scipy = "^1.8.0"
torch-scatter = "^2.0.9"
torch-cluster = "^1.6.0"
torch-sparse = "^0.6.13"
torch-geometric = "^2.0.4"
addict = "^2.4.0"
matplotlib = "^3.5.1"
PyYAML = "^6.0"
neuralogic = "^0.1.4"
spacy = "^3.2.3"
transformers = "^4.17.0"
nltk = "^3.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
